+++ frontend/src/Footer.tsx
@@ -8,16 +8,15 @@ import YouTubeIcon from "@material-ui/icons/YouTube";
 import ArrowUpwardRoundedIcon from "@material-ui/icons/ArrowUpwardRounded";
 import { Button } from "@material-ui/core";
 import {
-  ABOUT_US_TAB_PARAM,
+  TAB_PARAM,
   ABOUT_US_PAGE_LINK,
   EXPLORE_DATA_PAGE_LINK,
   DATA_CATALOG_PAGE_LINK,
   TERMS_OF_SERVICE_PAGE_LINK,
+  WHAT_IS_HEALTH_EQUITY_PAGE_LINK,
 } from "./utils/urlutils";
-import {
-  ABOUT_US_FAQ_TAB_INDEX,
-  ABOUT_US_CONTACT_TAB_INDEX,
-} from "./pages/AboutUs/AboutUsPage";
+import { ABOUT_US_CONTACT_TAB_INDEX } from "./pages/AboutUs/AboutUsPage";
+import { WIHE_FAQ_TAB_INDEX } from "./pages/WhatIsHealthEquity/WhatIsHealthEquityPage";
 
 function Footer() {
   return (
@@ -48,11 +47,11 @@ function Footer() {
               ["Downloads and Methods", DATA_CATALOG_PAGE_LINK],
               [
                 "FAQs",
-                `${ABOUT_US_PAGE_LINK}?${ABOUT_US_TAB_PARAM}=${ABOUT_US_FAQ_TAB_INDEX}`,
+                `${WHAT_IS_HEALTH_EQUITY_PAGE_LINK}?${TAB_PARAM}=${WIHE_FAQ_TAB_INDEX}`,
               ],
               [
                 "Contact Us",
-                `${ABOUT_US_PAGE_LINK}?${ABOUT_US_TAB_PARAM}=${ABOUT_US_CONTACT_TAB_INDEX}`,
+                `${ABOUT_US_PAGE_LINK}?${TAB_PARAM}=${ABOUT_US_CONTACT_TAB_INDEX}`,
               ],
               ["Terms of Use", `${TERMS_OF_SERVICE_PAGE_LINK}`],
             ].map(([label, link]) => (
